Restaurant Menu
A program that creates a restaurant menu.  The user inputs menu items
into the GUI (which was created with the help of an IDE).  After all
menu items are inputted, the user clicks “print,” and the menu is
printed to the screen.  In addition, a .txt file of the menu is saved
in the project folder.